// src/frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
// ...
export interface LineItem {
    rate: bigint;
    itemName: string;
    discount: bigint;
    quantity: bigint;
    taxRate: bigint;
    taxType: string;
}
export interface BusinessProfile {
    gstNumber: string;
    logo: ExternalBlob;
    plan: Plan;
    businessName: string;
    email: string;
    address: string;
    remainingInvoiceCredits: bigint;
    phone: string;
}
export interface Invoice {
    status: InvoiceStatus;
    lineItems: Array<LineItem>;
    client: Client;
    paymentStatus: PaymentStatus;
    dueDate: string;
    totalTax: bigint;
    grandTotal: bigint;
    invoiceDate: string;
    invoiceNumber: bigint;
    totalDiscount: bigint;
    subtotal: bigint;
}
// ...
export interface Client {
    clientId: bigint;
    gstNumber: string;
    name: string;
    email: string;
    address: string;
    companyName: string;
    phone: string;
}
export interface SavedItem {
    rate: bigint;
    description: string;
    itemCode: string;
    itemName: string;
    taxRate: bigint;
    taxType: Variant_igst_none_cgst_sgst;
}
export enum InvoiceStatus {
    paid = "paid",
    finalized = "finalized",
    overdue = "overdue",
    draft = "draft"
}
export enum PaymentStatus {
    paid = "paid",
    unpaid = "unpaid",
    partial = "partial"
}
export enum Plan {
    pro_single = "pro_single",
    pro_bundle = "pro_bundle",
    free = "free"
}